+++ src/main/java/net/tomasbot/matchday/db/converter/FFmpegMetadataConverter.java
@@ -21,7 +21,8 @@ package net.tomasbot.matchday.db.converter;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
-import net.tomasbot.matchday.plugin.io.ffmpeg.FFmpegMetadata;
+
+import net.tomasbot.ffmpeg_wrapper.metadata.FFmpegMetadata;
 import net.tomasbot.matchday.util.JsonParser;
 
 @Converter